+++ onnx/linear_test.go
@@ -4,15 +4,15 @@ import (
 	"fmt"
 	"testing"
 
-	"github.com/gomlx/gomlx/pkg/core/graph"
+	. "github.com/gomlx/gomlx/pkg/core/graph"
 	"github.com/gomlx/gomlx/pkg/core/graph/graphtest"
 	"github.com/gomlx/gomlx/pkg/core/tensors"
 	"github.com/gomlx/gomlx/pkg/ml/context"
 	"github.com/gomlx/gopjrt/dtypes"
 	"github.com/stretchr/testify/require"
-)
 
-import _ "github.com/gomlx/gomlx/backends/default"
+	_ "github.com/gomlx/gomlx/backends/default"
+)
 
 // TestEndToEnd based on the `linear_test.onnx` minimalistic model.
 // Only a couple of ops tested, but from end-to-end, including if changes can be saved
@@ -33,29 +33,35 @@ func TestEndToEnd(t *testing.T) {
 	ctx := context.New()
 	require.NoError(t, model.VariablesToContext(ctx))
 	for v := range ctx.IterVariables() {
-		fmt.Printf("\tVariable %q: %s\n", v.ScopeAndName(), v.Value())
+		value, err := v.Value()
+		require.NoError(t, err)
+		fmt.Printf("\tVariable %q: %s\n", v.ScopeAndName(), value)
 	}
 	vA := ctx.In(ModelScope).GetVariable("A")
 	require.NotNil(t, vA)
 	require.Equal(t, 1, vA.Shape().Rank())
 	require.Equal(t, 5, vA.Shape().Dim(0))
-	require.Equal(t, []float32{100, 10, 1, 0.1, 0.01}, tensors.CopyFlatData[float32](vA.Value()))
+	vAValue, err := vA.Value()
+	require.NoError(t, err)
+	require.Equal(t, []float32{100, 10, 1, 0.1, 0.01}, tensors.MustCopyFlatData[float32](vAValue))
 	vB := ctx.In(ModelScope).GetVariable("B")
 	require.NotNil(t, vB)
 	require.Equal(t, 0, vB.Shape().Rank())
-	require.Equal(t, float32(7000), tensors.ToScalar[float32](vB.Value()))
+	vBValue, err := vB.Value()
+	require.NoError(t, err)
+	require.Equal(t, float32(7000), tensors.ToScalar[float32](vBValue))
 
 	// Check conversion.
 	backend := graphtest.BuildTestBackend()
-	y := context.MustExecOnce(backend, ctx, func(ctx *context.Context, x *graph.Node) *graph.Node {
+	y := context.MustExecOnce(backend, ctx, func(ctx *context.Context, x *Node) *Node {
 		g := x.Graph()
-		outputs := model.CallGraph(ctx, g, map[string]*graph.Node{"X": x})
+		outputs := model.CallGraph(ctx, g, map[string]*Node{"X": x})
 		vB = ctx.In(ModelScope).GetVariable("B")
-		vB.SetValueGraph(graph.OnePlus(vB.ValueGraph(g)))
+		vB.SetValueGraph(OnePlus(vB.ValueGraph(g)))
 		return outputs[0]
 	}, [][]float32{{1, 2, 3, 4, 5}}) // BatchSize = 1
 	require.NoError(t, y.Shape().Check(dtypes.Float32, 1))
-	require.InDeltaSlice(t, []float32{7123.45}, tensors.CopyFlatData[float32](y), 0.1)
+	require.InDeltaSlice(t, []float32{7123.45}, tensors.MustCopyFlatData[float32](y), 0.1)
 
 	// Save change of variable "B" to the ONNX model.
 	require.NoError(t, model.ContextToONNX(ctx))